+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 # v3.5.0 <small><sup>06-08-2023</sup></small>
 
 - [Add] added [`setMissingDefaultsOnUpdate`](./v3.5.0/index.md#setmissingdefaultsonupdate) to schema options
+- [Update] updated types of `validate.isStringOk`
 
 # v3.4.0 <small><sup>04-08-2023</sup></small>
 
@@ -12,7 +13,7 @@ All notable changes to this project will be documented in this file.
 
 # v3.3.1 <small><sup>02-08-2023</sup></small>
 
-- [Update] improve typescript types
+- [Update] improved typescript types
 
 # v3.3.0 <small><sup>01-08-2023</sup></small>